test functions created + falling and stacking implemented

+++ SDK/test1/src/structures.h
@@ -21,8 +21,8 @@
  * */
 
 struct coords {
-	unsigned int x;
-	unsigned int y;
+	unsigned long int x;
+	unsigned long int y;
 };
 
 /*
